Trim whitespace and disallow internal whitespace for commodities
Commodities with preceding/trailing whitespace is not wanted, so the
symbol is trimmed before being added to the DB. Also these functions
returns the added commodity now since we are starting to manipulate them
on add.

Also disallow internal whitespace (e.g. "  U SD " is not OK while "  USD "
is OK and will be saved as "USD").