Manage migrations
For every change, there will be a new migration script that controls how
to upgrade the database from the previous version to the new one. This
script should ensure that there is no data corruption. When all migration
scripts have been run the database structure should be identical to running
`db.pgsql` straight away.

Versions are simply integers starting from 1. There is an internal table in
the database that keeps track of what versions have been deployed to the
database. This makes it easier to manage, plus it ensures that a migration
can't run twice.