Remove user and database creation from scripts
It doesn't really make sense, since these things varies from use case to
use case. For example there might be scenarios where several pg-ledger DBs
are to be run on the same server.

Moreover the security of the user accounts should be configured by the DBA
so they have the minimum permissions required by the specific application
that is using pg-ledger.