Add support for defining formulas
A "formula" consists of one to many "terms", each of which is linked to an
account (might add support later for also linking to other formulas in place
of accounts) and a sign which indicates whether the term should be added or
subtracted in the formula.

Accounts will be recursive by default. Might add support for customisation
of this later.

Example: Define net wealth as a formula with name "Net Wealth" and two terms:
  - Assets with "positive" set to `TRUE`, and
  - Liabilities with "positive" set to `FALSE`.